Ui shader addition and logger refactor
- Refactored prefix system into logger instances
- Added ui shader
- Added ui materials
- Added test render of ui quad
- Made vulkan objects more configurable to allow for different pipelines and shaders

+++ engine/src/core/c3d_string.cpp
@@ -14,9 +14,9 @@
 
 namespace C3D
 {
-	u64 StringLength(const string& str)
+	u32 StringLength(const string& str)
 	{
-		return str.size();
+		return static_cast<u32>(str.size());
 	}
 
 	bool Equals(const string& left, const string& right)
